switch to checking `WAYLAND_DISPLAY` from `XDG_SESSION_TYPE`
running with sudo previously caused a segfault. this was due to `XDG_SESSION_TYPE` being unset when run in the XWAYLAND process and attempting to do a string comparison on the null input. Now use `WAYLAND_DISPLAY` environment variable presence to more accurately test for wayland vs xwayland/x11 usage